+++ rapdev_rapid7/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG - rapdev_rapid7
 
+## 2.0.0 / 2022-1-21
+
+#### Feature Updates
+
+* [Added] Support for new Rapid7 Investigation API endpoint
+* [Added] Tracking of open/closed investigations via agent persistent cache
+* [Added] Datadog Events for newly opened/closed events
+* [Added] Dashboard to display open/closed event streams
+* [Changed] Changed existing log collection check to disabled by default
+
+
 ## 1.0.0 / 2021-09-15
 
 * [Added] initial release of the integration. Added all necessary files for marketplace merge.
+